+++ app/src/main/java/io/github/sds100/keymapper/constraints/ConstraintUiHelper.kt
@@ -1,26 +1,6 @@
 package io.github.sds100.keymapper.constraints
 
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Battery2Bar
-import androidx.compose.material.icons.outlined.BatteryChargingFull
-import androidx.compose.material.icons.outlined.BluetoothConnected
-import androidx.compose.material.icons.outlined.BluetoothDisabled
-import androidx.compose.material.icons.outlined.Call
-import androidx.compose.material.icons.outlined.CallEnd
-import androidx.compose.material.icons.outlined.FlashlightOff
-import androidx.compose.material.icons.outlined.FlashlightOn
-import androidx.compose.material.icons.outlined.Keyboard
-import androidx.compose.material.icons.outlined.Lock
-import androidx.compose.material.icons.outlined.LockOpen
-import androidx.compose.material.icons.outlined.MobileOff
-import androidx.compose.material.icons.outlined.PlayArrow
-import androidx.compose.material.icons.outlined.RingVolume
-import androidx.compose.material.icons.outlined.SignalWifiStatusbarNull
-import androidx.compose.material.icons.outlined.StayCurrentLandscape
-import androidx.compose.material.icons.outlined.StayCurrentPortrait
-import androidx.compose.material.icons.outlined.StopCircle
-import androidx.compose.material.icons.outlined.Wifi
-import androidx.compose.material.icons.outlined.WifiOff
 import androidx.compose.material.icons.rounded.Android
 import io.github.sds100.keymapper.R
 import io.github.sds100.keymapper.system.camera.CameraLensUtils
@@ -160,6 +140,8 @@ class ConstraintUiHelper(
         Constraint.PhoneRinging -> getString(R.string.constraint_phone_ringing)
         Constraint.Charging -> getString(R.string.constraint_charging)
         Constraint.Discharging -> getString(R.string.constraint_discharging)
+        Constraint.LockScreenShowing -> getString(R.string.constraint_lock_screen_showing)
+        Constraint.LockScreenNotShowing -> getString(R.string.constraint_lock_screen_not_showing)
     }
 
     fun getIcon(constraint: Constraint): ComposeIconInfo = when (constraint) {
@@ -175,62 +157,7 @@ class ConstraintUiHelper(
         is Constraint.AppNotPlayingMedia -> getAppIconInfo(constraint.packageName)
             ?: ComposeIconInfo.Vector(Icons.Rounded.Android)
 
-        Constraint.MediaPlaying -> ComposeIconInfo.Vector(Icons.Outlined.PlayArrow)
-
-        Constraint.NoMediaPlaying -> ComposeIconInfo.Vector(Icons.Outlined.StopCircle)
-
-        is Constraint.BtDeviceConnected -> ComposeIconInfo.Vector(Icons.Outlined.BluetoothConnected)
-
-        is Constraint.BtDeviceDisconnected -> ComposeIconInfo.Vector(Icons.Outlined.BluetoothDisabled)
-
-        is Constraint.OrientationCustom -> {
-            val resId = when (constraint.orientation) {
-                Orientation.ORIENTATION_0 -> Icons.Outlined.StayCurrentPortrait
-                Orientation.ORIENTATION_90 -> Icons.Outlined.StayCurrentLandscape
-                Orientation.ORIENTATION_180 -> Icons.Outlined.StayCurrentPortrait
-                Orientation.ORIENTATION_270 -> Icons.Outlined.StayCurrentLandscape
-            }
-
-            ComposeIconInfo.Vector(resId)
-        }
-
-        Constraint.OrientationLandscape -> ComposeIconInfo.Vector(Icons.Outlined.StayCurrentLandscape)
-
-        Constraint.OrientationPortrait -> ComposeIconInfo.Vector(Icons.Outlined.StayCurrentPortrait)
-
-        Constraint.ScreenOff -> ComposeIconInfo.Vector(Icons.Outlined.MobileOff)
-
-        Constraint.ScreenOn -> ComposeIconInfo.Vector(Icons.Outlined.StayCurrentPortrait)
-
-        is Constraint.FlashlightOff -> ComposeIconInfo.Vector(Icons.Outlined.FlashlightOff)
-
-        is Constraint.FlashlightOn -> ComposeIconInfo.Vector(Icons.Outlined.FlashlightOn)
-
-        is Constraint.WifiConnected -> ComposeIconInfo.Vector(Icons.Outlined.Wifi)
-
-        is Constraint.WifiDisconnected -> ComposeIconInfo.Vector(Icons.Outlined.SignalWifiStatusbarNull)
-
-        Constraint.WifiOff -> ComposeIconInfo.Vector(Icons.Outlined.WifiOff)
-
-        Constraint.WifiOn -> ComposeIconInfo.Vector(Icons.Outlined.Wifi)
-
-        is Constraint.ImeChosen -> ComposeIconInfo.Vector(Icons.Outlined.Keyboard)
-
-        is Constraint.ImeNotChosen -> ComposeIconInfo.Vector(Icons.Outlined.Keyboard)
-
-        Constraint.DeviceIsLocked -> ComposeIconInfo.Vector(Icons.Outlined.Lock)
-
-        Constraint.DeviceIsUnlocked -> ComposeIconInfo.Vector(Icons.Outlined.LockOpen)
-
-        Constraint.InPhoneCall -> ComposeIconInfo.Vector(Icons.Outlined.Call)
-
-        Constraint.NotInPhoneCall -> ComposeIconInfo.Vector(Icons.Outlined.CallEnd)
-
-        Constraint.PhoneRinging -> ComposeIconInfo.Vector(Icons.Outlined.RingVolume)
-
-        Constraint.Charging -> ComposeIconInfo.Vector(Icons.Outlined.BatteryChargingFull)
-
-        Constraint.Discharging -> ComposeIconInfo.Vector(Icons.Outlined.Battery2Bar)
+        else -> ConstraintUtils.getIcon(constraint.id)
     }
 
     private fun getAppIconInfo(packageName: String): ComposeIconInfo? = getAppIcon(packageName).handle(