+++ app/src/main/java/io/github/sds100/keymapper/constraints/DetectConstraintsUseCase.kt
@@ -73,6 +73,9 @@ class DetectConstraintsUseCaseImpl(
                 emptyFlow()
             }
 
+            ConstraintDependency.LOCK_SCREEN_SHOWING ->
+                lockScreenAdapter.isLockScreenShowingFlow().map { dependency }
+
             ConstraintDependency.PHONE_STATE -> phoneAdapter.callStateFlow.map { dependency }
             ConstraintDependency.CHARGING_STATE -> powerAdapter.isCharging.map { dependency }
         }