+++ app/src/main/java/io/github/sds100/keymapper/sorting/comparators/KeyMapConstraintsComparator.kt
@@ -123,6 +123,8 @@ class KeyMapConstraintsComparator(
 
             Constraint.WifiOff -> Success("")
             Constraint.WifiOn -> Success("")
+            Constraint.LockScreenNotShowing -> Success("")
+            Constraint.LockScreenShowing -> Success("")
         }
     }
 }