+++ app/src/main/java/io/github/sds100/keymapper/system/lock/LockScreenAdapter.kt
@@ -16,4 +16,7 @@ interface LockScreenAdapter {
 
     @RequiresApi(Build.VERSION_CODES.LOLLIPOP_MR1)
     fun isLockedFlow(): Flow<Boolean>
+
+    fun isLockScreenShowing(): Boolean
+    fun isLockScreenShowingFlow(): Flow<Boolean>
 }